renamed filterbutton component
